fix rerender bug
before, when clicking on a related category on a categoy page, the properties on the page would not update - they would still be the same as before. the load functions and the data on the client look fine. but svelte did not rerender. by adding this key, it does. there should be a more elegant solution!